Architecture_1/src/function/src_upload_to_bigquery/requirements.txt
google.cloud.bigquery==3.*
functions_framework==3.*